requirements.txt
argparse~=1.4.0
matplotlib~=3.3.4
mininet~=2.3.0
Mini-NDN~=0.5.0
pandas~=1.2.3
python-ndn~=0.3a1
seaborn~=0.11.1
